passage norminette

+++ Libft/ft_strsplit.c
@@ -1,21 +1,21 @@
-/*  ************************************************************************** */
+/* ************************************************************************** */
 /*                                                                            */
 /*                                                        :::      ::::::::   */
 /*   ft_strsplit.c                                      :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
 /*   By: cghanime <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
-/*   Created: 2018/11/14 16:55:38 by cghanime          #+#    #+#             */
-/*   Updated: 2018/11/20 12:15:58 by cghanime         ###   ########.fr       */
+/*   Created: 2018/11/20 19:14:30 by cghanime          #+#    #+#             */
+/*   Updated: 2018/11/20 19:30:34 by cghanime         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-static int		ft_wordcount(char const *s, char c)
+static	int		ft_wordcount(char const *s, char c)
 {
-	size_t			i;
-	unsigned int	wordcount;
+	size_t		i;
+	size_t		wordcount;
 
 	i = 0;
 	wordcount = 0;
@@ -31,11 +31,11 @@ static int		ft_wordcount(char const *s, char c)
 	return (wordcount);
 }
 
-static char **ft_filling(char const *s, char c, char **tab)
+static	char	**ft_filling(char const *s, char c, char **tab)
 {
-	size_t begin;
-	size_t end;
-	size_t i;
+	size_t		begin;
+	size_t		end;
+	size_t		i;
 
 	end = 0;
 	i = 0;
@@ -54,11 +54,11 @@ static char **ft_filling(char const *s, char c, char **tab)
 	return (tab);
 }
 
-char	**ft_strsplit(char const *s, char c)
+char			**ft_strsplit(char const *s, char c)
 {
-	char **str;
+	char		**str;
 
 	if (!(str = (char**)malloc(sizeof(char*) * (ft_wordcount(s, c) + 1))))
 		return (NULL);
-	return(ft_filling(s, c, str));
+	return (ft_filling(s, c, str));
 }
